Convolutions

A convolution of two probability distributions is the probability distribution of the sum of two independent random variables that are distributed according to these distributions.

The convolution of two distributions can be constructed with convolve.

Distributions.convolve — Function

convolve(d1::Distribution, d2::Distribution)

Convolve two distributions and return the distribution corresponding to the sum of independent random variables drawn from the underlying distributions.

Currently, the function is only defined in cases where the convolution has a closed form. More precisely, the function is defined if the distributions of d1 and d2 are the same and one of

Bernoulli

Binomial

NegativeBinomial

Geometric

Poisson

DiscreteNonParametric

Normal

Cauchy

Chisq

Exponential

Gamma

MvNormal

External links: List of convolutions of probability distributions on Wikipedia