Matrix-variate Distributions

Matrix-variate Distributions

Matrix-variate distributions are the distributions whose variate forms are Matrixvariate (i.e each sample is a matrix). Abstract types for matrix-variate distributions:

Common Interface

Both distributions implement the same set of methods:

Base.size — Method.

size(d::MatrixDistribution)

Return the size of each sample from distribution d.

source

Base.length — Method.

length(d::MatrixDistribution)

The length (i.e number of elements) of each sample from the distribution d.

source

Statistics.mean — Method.

mean(d::MatrixDistribution)

Return the mean matrix of d.

source

Distributions.pdf — Method.

pdf(d::MatrixDistribution, x::AbstractArray)

Compute the probability density at the input matrix x.

source

Distributions.logpdf — Method.

logpdf(d::MatrixDistribution, AbstractMatrix)

Compute the logarithm of the probability density at the input matrix x.

source

Distributions._rand! — Method.

_rand!(::AbstractRNG, ::MatrixDistribution, A::AbstractMatrix)

Sample the matrix distribution and store the result in A. Must be implemented by matrix-variate distributions.

source

Distributions

Distributions.Wishart — Type.

Wishart(nu, S)

The Wishart distribution is a multidimensional generalization of the Chi-square distribution, which is characterized by a degree of freedom ν, and a base matrix S.

source

Distributions.InverseWishart — Type.

InverseWishart(nu, P)

The Inverse Wishart distribution is usually used as the conjugate prior for the covariance matrix of a multivariate normal distribution, which is characterized by a degree of freedom ν, and a base matrix Φ.

source

Internal Methods (for creating your own matrix-variate distributions)

Distributions._logpdf — Method.

_logpdf(d::MatrixDistribution, x::AbstractArray)

Evaluate logarithm of pdf value for a given sample x. This function need not perform dimension checking.

source