Matrix-variate Distributions

Matrix-variate Distributions

Matrix-variate distributions are the distributions whose variate forms are Matrixvariate (i.e each sample is a matrix). Abstract types for matrix-variate distributions:

Common Interface

All distributions implement the same set of methods:

Base.size — Method.

size(d::MatrixDistribution)

Return the size of each sample from distribution d.

source

Base.length — Method.

length(d::MatrixDistribution)

The length (i.e number of elements) of each sample from the distribution d.

source

Statistics.mean — Method.

mean(d::MatrixDistribution)

Return the mean matrix of d.

source

Distributions.pdf — Method.

pdf(d::MatrixDistribution, x::AbstractArray)

Compute the probability density at the input matrix x.

source

Distributions.logpdf — Method.

logpdf(d::MatrixDistribution, AbstractMatrix)

Compute the logarithm of the probability density at the input matrix x.

source

Distributions._rand! — Method.

_rand!(::AbstractRNG, ::MatrixDistribution, A::AbstractMatrix)

Sample the matrix distribution and store the result in A. Must be implemented by matrix-variate distributions.

source

Base.vec — Method.

vec(d::MatrixDistribution)

If known, returns a MultivariateDistribution instance representing the distribution of vec(X), where X is a random matrix with distribution d.

source

Distributions

Distributions.Wishart — Type.

Wishart(nu, S)

The Wishart distribution is a multidimensional generalization of the Chi-square distribution, which is characterized by a degree of freedom ν, and a base matrix S.

source

Distributions.InverseWishart — Type.

InverseWishart(nu, P)

The Inverse Wishart distribution is usually used as the conjugate prior for the covariance matrix of a multivariate normal distribution, which is characterized by a degree of freedom ν, and a base matrix Φ.

source

Distributions.MatrixNormal — Type.

MatrixNormal(M, U, V)

M::AbstractMatrix  n x p mean
U::PDMat           n x n row covariance
V::PDMat           p x p column covariance

The matrix normal distribution generalizes the multivariate normal distribution to $n\times p$ real matrices $\mathbf{X}$. If $\mathbf{X}\sim MN(\mathbf{M}, \mathbf{U}, \mathbf{V})$, then its probability density function is

\[f(\mathbf{X};\mathbf{M}, \mathbf{U}, \mathbf{V}) = \frac{\exp\left( -\frac{1}{2} \, \mathrm{tr}\left[ \mathbf{V}^{-1} (\mathbf{X} - \mathbf{M})^{\rm{T}} \mathbf{U}^{-1} (\mathbf{X} - \mathbf{M}) \right] \right)}{(2\pi)^{np/2} |\mathbf{V}|^{n/2} |\mathbf{U}|^{p/2}}.\]

$\mathbf{X}\sim MN(\mathbf{M},\mathbf{U},\mathbf{V})$ if and only if $\text{vec}(\mathbf{X})\sim N(\text{vec}(\mathbf{M}),\mathbf{V}\otimes\mathbf{U})$.

source

Distributions.MatrixTDist — Type.

MatrixTDist(ν, M, Σ, Ω)

ν::Real            positive degrees of freedom
M::AbstractMatrix  n x p location
Σ::PDMat           n x n scale
Ω::PDMat           p x p scale

The matrix t-Distribution generalizes the multivariate t-Distribution to $n\times p$ real matrices $\mathbf{X}$. If $\mathbf{X}\sim MT(\nu,\mathbf{M},\boldsymbol{\Sigma}, \boldsymbol{\Omega})$, then its probability density function is

\[f(\mathbf{X} ; \nu,\mathbf{M},\boldsymbol{\Sigma}, \boldsymbol{\Omega}) = c_0 \left|\mathbf{I}_n + \boldsymbol{\Sigma}^{-1}(\mathbf{X} - \mathbf{M})\boldsymbol{\Omega}^{-1}(\mathbf{X}-\mathbf{M})'\right|^{-\frac{\nu+n+p-1}{2}},\]

where

\[c_0=\frac{\Gamma_p\left(\frac{\nu+n+p-1}{2}\right)}{(\pi)^\frac{np}{2} \Gamma_p\left(\frac{\nu+p-1}{2}\right)} |\boldsymbol{\Omega}|^{-\frac{n}{2}} |\boldsymbol{\Sigma}|^{-\frac{p}{2}}.\]

If the joint distribution $p(\mathbf{S},\mathbf{X})=p(\mathbf{S})p(\mathbf{X}|\mathbf{S})$ is given by

\[\begin{align*} \mathbf{S}&\sim IW(\nu + n - 1, \boldsymbol{\Sigma})\\ \mathbf{X}|\mathbf{S}&\sim MN(\mathbf{M}, \mathbf{S}, \boldsymbol{\Omega}), \end{align*}\]

then the marginal distribution of $\mathbf{X}$ is $MT(\nu,\mathbf{M},\boldsymbol{\Sigma},\boldsymbol{\Omega})$.

source

Internal Methods (for creating your own matrix-variate distributions)

Distributions._logpdf — Method.

_logpdf(d::MatrixDistribution, x::AbstractArray)

Evaluate logarithm of pdf value for a given sample x. This function need not perform dimension checking.

source